Fix metric tulip miscalculation

+++ segmmodules/segmtulip.cpp
@@ -585,8 +585,8 @@ AnalysisResult SegmentTulip::run(Communicator *comm, ShapeGraph &map, bool) {
                             case RadiusType::METRIC:
                                 while (rbin != static_cast<int>(nradii) &&
                                        radius[static_cast<size_t>(rbin)] != -1 &&
-                                       static_cast<double>(lineindex.metricdepth + seglength) *
-                                               0.5 >
+                                       static_cast<double>(lineindex.metricdepth) +
+                                               (static_cast<double>(seglength) * 0.5) >
                                            radius[static_cast<size_t>(rbin)]) {
                                     rbin++;
                                 }
@@ -654,8 +654,8 @@ AnalysisResult SegmentTulip::run(Communicator *comm, ShapeGraph &map, bool) {
                             case RadiusType::METRIC:
                                 while (rbin != static_cast<int>(nradii) &&
                                        radius[static_cast<size_t>(rbin)] != -1 &&
-                                       static_cast<double>(lineindex.metricdepth + seglength) *
-                                               0.5 >
+                                       static_cast<double>(lineindex.metricdepth) +
+                                               (static_cast<double>(seglength) * 0.5) >
                                            radius[static_cast<size_t>(rbin)]) {
                                     rbin++;
                                 }
